Correcciones de tipos unificados (ulong - long). Uso la endidad del Dominio. Correciones varias.

+++ src/SGVO.Infrastructure/Commands/EliminarUnidadOrganizativaCommandHandler.cs
@@ -24,7 +24,7 @@ public class EliminarUnidadOrganizativaCommandHandler : ICommandHandler<Eliminar
         CancellationToken cancellationToken)
     {
         var entity = await _dbContext.UnidadesOrganizativas
-            .FirstOrDefaultAsync(e => e.Id == (ulong)command.Id, cancellationToken);
+            .FirstOrDefaultAsync(e => e.Id == command.Id, cancellationToken);
 
         if (entity is null)
             return Result<Unit>.Failure(
